test: centralize record test helpers and refactor generators
- Move `get_test_record_batch` from `src/lib.rs` test module to `src/record/test
.rs` under `#[cfg(all(test, feature = "tokio"))]`
- Update imports in `src/ondisk/sstable.rs` tests to use `crate::record::test::{
get_test_record_batch, Test}`
- Introduce `test_items<I>(range: I) -> impl Iterator<Item = Test>` in `src/reco
rd/test.rs`
- Refactor all call sites in `src/lib.rs` tests to use `test_items(0u32..N)` ins
tead of local `test_items()` and slicing
- Remove old helpers from `src/lib.rs` and clean unused imports
- Clippy clean and all tests pass
