+++ src/lib.rs
@@ -1116,234 +1116,34 @@ pub type ParquetLru = Arc<dyn DynLruCache<FileId> + Send + Sync>;
 pub(crate) mod tests {
     use std::{
         collections::{BTreeMap, Bound},
-        mem,
         sync::Arc,
     };
 
-    use arrow::{
-        array::{Array, AsArray, RecordBatch},
-        datatypes::{Schema, UInt32Type},
-    };
     use flume::{bounded, Receiver};
-    use fusio::{disk::TokioFs, path::Path, DynFs, SeqRead, Write};
+    use fusio::{disk::TokioFs, path::Path, DynFs};
     use fusio_dispatch::FsOptions;
-    use fusio_log::{Decode, Encode};
     use futures::StreamExt;
-    use parquet::arrow::ProjectionMask;
     use parquet_lru::NoCache;
     use tempfile::TempDir;
     use tracing::error;
 
+    pub use crate::record::test::{Test, TestRef};
     use crate::{
         compaction::{error::CompactionError, leveled::LeveledCompactor, CompactTask, Compactor},
         context::Context,
-        executor::{tokio::TokioExecutor, Executor, RwLock},
+        executor::{tokio::TokioExecutor, Executor},
         fs::{generate_file_id, manager::StoreManager},
         inmem::{immutable::tests::TestSchema, mutable::MutableMemTable},
         manifest::ManifestStorageError,
         record::{
             dynamic::test::{test_dyn_item_schema, test_dyn_items},
-            option::OptionRecordRef,
-            DynRecord, Key, KeyRef, RecordRef, Schema as RecordSchema, Value, ValueRef,
+            DynRecord, KeyRef, Schema as RecordSchema, Value, ValueRef,
         },
         trigger::{TriggerFactory, TriggerType},
         version::{cleaner::Cleaner, set::tests::build_version_set, Version},
         wal::log::LogType,
         CompactionOption, DbError, DbOption, Projection, Record, DB,
     };
-
-    #[derive(Debug, PartialEq, Eq, Clone)]
-    pub struct Test {
-        pub vstring: String,
-        pub vu32: u32,
-        pub vbool: Option<bool>,
-    }
-
-    impl Decode for Test {
-        async fn decode<R>(reader: &mut R) -> Result<Self, fusio::Error>
-        where
-            R: SeqRead,
-        {
-            let vstring = String::decode(reader).await?;
-            let vu32 = Option::<u32>::decode(reader).await?.unwrap();
-            let vbool = Option::<bool>::decode(reader).await?;
-
-            Ok(Self {
-                vstring,
-                vu32,
-                vbool,
-            })
-        }
-    }
-
-    impl Record for Test {
-        type Schema = TestSchema;
-
-        type Ref<'r>
-            = TestRef<'r>
-        where
-            Self: 'r;
-
-        fn key(&self) -> &str {
-            &self.vstring
-        }
-
-        fn as_record_ref(&self) -> Self::Ref<'_> {
-            TestRef {
-                vstring: &self.vstring,
-                vu32: Some(self.vu32),
-                vbool: self.vbool,
-            }
-        }
-
-        fn size(&self) -> usize {
-            let string_size = self.vstring.len();
-            let u32_size = mem::size_of::<u32>();
-            let bool_size = self.vbool.map_or(0, |_| mem::size_of::<bool>());
-            string_size + u32_size + bool_size
-        }
-    }
-
-    #[derive(Debug, PartialEq, Eq, Clone, Copy)]
-    pub struct TestRef<'r> {
-        pub vstring: &'r str,
-        pub vu32: Option<u32>,
-        pub vbool: Option<bool>,
-    }
-
-    impl Encode for TestRef<'_> {
-        async fn encode<W>(&self, writer: &mut W) -> Result<(), fusio::Error>
-        where
-            W: Write,
-        {
-            self.vstring.encode(writer).await?;
-            self.vu32.encode(writer).await?;
-            self.vbool.encode(writer).await?;
-
-            Ok(())
-        }
-
-        fn size(&self) -> usize {
-            self.vstring.size() + self.vu32.size() + self.vbool.size()
-        }
-    }
-
-    impl<'r> RecordRef<'r> for TestRef<'r> {
-        type Record = Test;
-
-        fn key(self) -> <<<Self::Record as Record>::Schema as RecordSchema>::Key as Key>::Ref<'r> {
-            self.vstring
-        }
-
-        fn projection(&mut self, projection_mask: &ProjectionMask) {
-            if !projection_mask.leaf_included(3) {
-                self.vu32 = None;
-            }
-            if !projection_mask.leaf_included(4) {
-                self.vbool = None;
-            }
-        }
-
-        fn from_record_batch(
-            record_batch: &'r RecordBatch,
-            offset: usize,
-            projection_mask: &'r ProjectionMask,
-            _: &Arc<Schema>,
-        ) -> OptionRecordRef<'r, Self> {
-            let mut column_i = 2;
-            let null = record_batch.column(0).as_boolean().value(offset);
-
-            let ts = record_batch
-                .column(1)
-                .as_primitive::<UInt32Type>()
-                .value(offset)
-                .into();
-
-            let vstring = record_batch
-                .column(column_i)
-                .as_string::<i32>()
-                .value(offset);
-            column_i += 1;
-
-            let mut vu32 = None;
-
-            if projection_mask.leaf_included(3) {
-                vu32 = Some(
-                    record_batch
-                        .column(column_i)
-                        .as_primitive::<UInt32Type>()
-                        .value(offset),
-                );
-                column_i += 1;
-            }
-
-            let mut vbool = None;
-
-            if projection_mask.leaf_included(4) {
-                let vbool_array = record_batch.column(column_i).as_boolean();
-
-                if !vbool_array.is_null(offset) {
-                    vbool = Some(vbool_array.value(offset));
-                }
-            }
-
-            let record = TestRef {
-                vstring,
-                vu32,
-                vbool,
-            };
-            OptionRecordRef::new(ts, record, null)
-        }
-    }
-
-    pub(crate) async fn get_test_record_batch<E: Executor + Send + Sync + 'static>(
-        option: DbOption,
-        executor: E,
-    ) -> RecordBatch {
-        let db: DB<Test, E> = DB::new(option.clone(), executor, TestSchema {})
-            .await
-            .unwrap();
-        let base_fs = db.ctx.manager.base_fs();
-
-        db.write(
-            Test {
-                vstring: "hello".to_string(),
-                vu32: 12,
-                vbool: Some(true),
-            },
-            1.into(),
-        )
-        .await
-        .unwrap();
-        db.write(
-            Test {
-                vstring: "world".to_string(),
-                vu32: 12,
-                vbool: None,
-            },
-            1.into(),
-        )
-        .await
-        .unwrap();
-
-        let mut schema = db.mem_storage.write().await;
-
-        let trigger = schema.trigger.clone();
-        let mutable = mem::replace(
-            &mut schema.mutable,
-            MutableMemTable::new(&option, trigger, base_fs.clone(), Arc::new(TestSchema {}))
-                .await
-                .unwrap(),
-        );
-
-        mutable
-            .into_immutable()
-            .await
-            .unwrap()
-            .1
-            .as_record_batch()
-            .clone()
-    }
 
     pub(crate) async fn build_schema(
         option: Arc<DbOption>,
@@ -1462,6 +1262,8 @@ pub(crate) mod tests {
         ))
     }
 
+    use crate::record::test::test_items;
+
     pub(crate) async fn build_db<R, E>(
         option: Arc<DbOption>,
         compaction_rx: Receiver<CompactTask>,
@@ -1540,18 +1342,6 @@ pub(crate) mod tests {
         })
     }
 
-    fn test_items() -> Vec<Test> {
-        let mut items = Vec::new();
-        for i in 0..32 {
-            items.push(Test {
-                vstring: i.to_string(),
-                vu32: i,
-                vbool: Some(true),
-            });
-        }
-        items
-    }
-
     #[tokio::test(flavor = "multi_thread")]
     async fn read_from_disk() {
         let temp_dir = TempDir::new().unwrap();
@@ -1575,7 +1365,7 @@ pub(crate) mod tests {
             .await
             .unwrap();
 
-        for (i, item) in test_items().into_iter().enumerate() {
+        for (i, item) in test_items(0u32..32).enumerate() {
             db.write(item, 0.into()).await.unwrap();
             if i % 5 == 0 {
                 db.flush().await.unwrap();
@@ -1616,11 +1406,11 @@ pub(crate) mod tests {
             .await
             .unwrap();
 
-        for item in &test_items()[0..10] {
+        for item in test_items(0u32..10) {
             db.write(item.clone(), 0.into()).await.unwrap();
         }
         db.flush().await.unwrap();
-        for item in &test_items()[10..20] {
+        for item in test_items(10u32..20) {
             db.write(item.clone(), 0.into()).await.unwrap();
         }
 
@@ -1674,7 +1464,7 @@ pub(crate) mod tests {
                 .await
                 .unwrap();
 
-            for item in &test_items()[0..10] {
+            for item in test_items(0u32..10) {
                 db.insert(item.clone()).await.unwrap();
             }
             // flush to s3
@@ -1701,7 +1491,7 @@ pub(crate) mod tests {
                     .await
                     .unwrap();
             let mut sort_items = BTreeMap::new();
-            for item in test_items()[0..10].iter() {
+            for item in test_items(0u32..10) {
                 sort_items.insert(item.vstring.clone(), item.clone());
             }
             let tx = db.transaction().await;
@@ -1748,7 +1538,7 @@ pub(crate) mod tests {
             option: option.clone(),
         };
 
-        for (i, item) in test_items().into_iter().enumerate() {
+        for (i, item) in test_items(0u32..32).enumerate() {
             mem_storage
                 .write(LogType::Full, item, (i as u32).into())
                 .await
@@ -1766,7 +1556,7 @@ pub(crate) mod tests {
         .unwrap();
 
         let mut sort_items = BTreeMap::new();
-        for item in test_items() {
+        for item in test_items(0u32..32) {
             sort_items.insert(item.vstring.clone(), item);
         }
         {
@@ -1885,7 +1675,7 @@ pub(crate) mod tests {
             .await
             .unwrap();
 
-        for (idx, item) in test_items().into_iter().enumerate() {
+        for (idx, item) in test_items(0u32..32).enumerate() {
             if idx % 2 == 0 {
                 db.write(item, 0.into()).await.unwrap();
             } else {
@@ -2208,16 +1998,5 @@ pub(crate) mod tests {
                 i += 1
             }
         }
-    }
-
-    #[test]
-    fn build_test() {
-        let t = trybuild::TestCases::new();
-        t.pass("tests/success/*.rs");
-    }
-    #[test]
-    fn fail_build_test() {
-        let t = trybuild::TestCases::new();
-        t.compile_fail("tests/fail/*.rs");
     }
 }
